move code outside __init__
